Move unit test to cover all features
